feat: refactor two-factor authentication plugin by modularizing passkey and TOTP functionalities
